Simplify queue processing function

+++ src/Template.Functions/Shared/QueueConstants.cs
@@ -8,5 +8,5 @@ public static class QueueConstants {
   /// This value is actually controlled in host.json and defaults to 5 when not set. Please ensure this const value
   /// stays in-sync with the value assigned in host.json
   /// </remarks>
-  public const int MaxDequeueCount = 5;
+  public const long MaxDequeueCount = 5;
 }
